Microservices/Complex_Microservices/end_auction/endAuction_requirements.txt
flask_sqlalchemy
flask_cors
pika
pyodbc
requests==2.31.0